Allow all listgroup empty messages to be translated
