+++ blocks/listgroup-custom/listgroup-custom.js
@@ -16,6 +16,8 @@ import {
   writeImagePropertyInList,
 } from '../../scripts/jmp.js';
 
+import { getEmptyResultsMessage } from '../../scripts/listgroup.js';
+
 import { loadScript } from '../../scripts/aem.js';
 import { createTag } from '../../scripts/helper.js';
 
@@ -60,23 +62,6 @@ function processDate(dateProperty, prop, item) {
     span = `<span class="${prop}">${item[prop]}</span>`;
   }
   return span;
-}
-
-async function getEmptyResultsMessage(emptyResultString) {
-  if (!emptyResultString) {
-    return undefined;
-  }
-
-  if (emptyResultString.includes('.json')) {
-    const pageLanguage = getLanguage();
-    const data = await getJsonFromUrl(emptyResultString);
-    const { data: translations } = data[pageLanguage];
-    console.log(translations[0]);
-    return translations[0].emptyResultsMessage;
-  } else {
-    // use the string as the empty results message. It won't translate.
-    return emptyResultString;
-  }
 }
 
 async function getTagTranslations() {
