+++ blocks/listgroup-events-tabs/listgroup-events-tabs.js
@@ -13,6 +13,8 @@ import {
   pageOrFilter,
 } from '../../scripts/jmp.js';
 
+import { getEmptyResultsMessage } from '../../scripts/listgroup.js';
+
 /*
  * Apply where a property is not empty.
  */
@@ -68,7 +70,8 @@ export default async function decorate(block) {
   const optionsObject = getBlockPropertiesList(block, 'options');
   const tabs = getBlockPropertiesList(block, 'tabs');
   const startingFolder = getBlockProperty(block, 'startingFolder');
-  const emptyResultsMessage = getBlockProperty(block, 'emptyResultsMessage');
+  const emptyResultsMessageConfig = getBlockProperty(block, 'emptyResultsMessage');
+  const emptyResultsMessage = await getEmptyResultsMessage(emptyResultsMessageConfig);
   const filterOptions = getListFilterOptions(block, propertyNames);
 
   // If startingFolder is not null, then apply page location filter FIRST.
